Changed build system to create three different products:
- minimal: This product is only a sceleton to plugin all features. It
will offer only an base Eclipse functionality for perspectives, plugins
and about. This product is not shipped. It is used to check the plugins
and features to be loaded correctly. It is used when no other running
Eclipse application is there like Eclipse IDE.
- base: This product contains minimal and additionally all common stuff.
This will be shipped per default.
- full: This product contains everything which is needed for testing.
This will normally not be shipped.